Refactors Harbor problems code and adds v1 Jest tests
