+++ node-packages/commons/src/api.ts
@@ -164,6 +164,7 @@ fragment on Project {
   name
   gitUrl
   privateKey
+  problemsUi
 }
 `);
