+++ services/webhooks2tasks/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  roots: ['./src'],
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest'
+  }
+};
