+++ services/webhooks2tasks/src/handlers/problems/harborExceptions.ts
@@ -1,0 +1,14 @@
+export class ProblemsHarborConnectionError extends Error {
+    constructor(message) {
+      super(message);
+      this.name = 'problems-harborConnectionError';
+    }
+  }
+
+export class ProblemsInvalidWebhookData extends Error {
+    constructor(message) {
+      super(message);
+      this.name = 'problems-invalidWebhookData';
+    }
+}
+
