+++ services/webhooks2tasks/src/handlers/problems/harborScanningCompleted.ts
@@ -1,11 +1,17 @@
 // @flow
 
+//TODO:
+// 1 - replace console.logs with actual lagoon log errors
+// 2 - add Jest tests for Harbor v1 and v2 for validateAndTransformIncomingWebhookdata - this is where the rubber hits the road.
+// 3 - add Jest tests for extractVulnerabilities
+
 import { sendToLagoonLogs } from '@lagoon/commons/dist/logs';
 import {
   getVulnerabilitiesPayloadFromHarbor,
 } from '@lagoon/commons/dist/harborApi';
 import * as R from 'ramda';
 import uuid4 from 'uuid4';
+import {extractVulnerabilities, matchRepositoryAgainstPatterns, generateError, validateAndTransformIncomingWebhookdata} from './harborHelpers'
 
 import {
   getProjectByName,
@@ -18,21 +24,14 @@ const PROBLEMS_HARBOR_FILTER_FLAG = process.env.PROBLEMS_HARBOR_FILTER_FLAG || n
 
 const HARBOR_WEBHOOK_SUCCESSFUL_SCAN = "Success";
 
-const DEFAULT_REPO_DETAILS_REGEX = "^(?<lagoonProjectName>.+)\/(?<lagoonEnvironmentName>.+)\/(?<lagoonServiceName>.+)$";
-
-const DEFAULT_REPO_DETAILS_MATCHER = {
-  defaultProjectName: "",
-  defaultEnvironmentName: "",
-  defaultServiceName: "",
-  regex: DEFAULT_REPO_DETAILS_REGEX,
-};
-
  export async function harborScanningCompleted(
   WebhookRequestData,
   channelWrapperWebhooks
 ) {
   const { webhooktype, event, uuid, body } = WebhookRequestData;
   const HARBOR_WEBHOOK_SUCCESSFUL_SCAN = "Success";
+
+  let harborScanPatternMatchers = await getProblemHarborScanMatches();
 
   try {
     let {
@@ -43,7 +42,7 @@ const DEFAULT_REPO_DETAILS_MATCHER = {
       lagoonEnvironmentName,
       lagoonServiceName,
       harborScanId,
-    } = await validateAndTransformIncomingWebhookdata(body);
+    } = await validateAndTransformIncomingWebhookdata(harborScanPatternMatchers.allProblemHarborScanMatchers, body);
 
     if(scanOverview.scan_status !== HARBOR_WEBHOOK_SUCCESSFUL_SCAN) {
       sendToLagoonLogs(
@@ -120,86 +119,7 @@ const DEFAULT_REPO_DETAILS_MATCHER = {
   }
 }
 
-/**
- * This function will take an incoming Harbor webhook and decompose it
- * into a more useable format
- *
- * @param {*} rawData
- */
-const validateAndTransformIncomingWebhookdata = async (rawData) => {
-  let { resources, repository } = rawData.event_data;
 
-  if (!repository.repo_full_name) {
-    throw generateError(
-      'InvalidHarborInput',
-      'Unable to find repo_full_name in body.event_data.repository'
-    );
-  }
-
-  // scan_overview is tricky because the property doesn't have an obvious name.
-  // We convert it to an array of objects with the old property as a member
-  let scanOverviewArray = R.toPairs(resources[0].scan_overview).map((e) => {
-    let obj = e[1];
-    obj.scan_key = e[0];
-    return obj;
-  });
-
-  let harborScanPatternMatchers = await getProblemHarborScanMatches();
-
-  let {
-    lagoonProjectName,
-    lagoonEnvironmentName,
-    lagoonServiceName,
-   } = matchRepositoryAgainstPatterns(repository.repo_full_name, harborScanPatternMatchers.allProblemHarborScanMatchers);
-
-  return {
-    resources,
-    repository,
-    scanOverview: scanOverviewArray.pop(),
-    lagoonProjectName,
-    lagoonEnvironmentName,
-    lagoonServiceName,
-    harborScanId: repository.repo_full_name,
-  };
-};
-
-const generateError = (name, message) => {
-  let e = new Error(message);
-  e.name = name;
-  return e;
-};
-
-const matchRepositoryAgainstPatterns = (repoFullName, matchPatterns = []) => {
-  const matchingRes = matchPatterns.filter((e) => generateRegex(e.regex).test(repoFullName));
-
-  if(matchingRes.length > 1) {
-    const stringifyMatchingRes = matchingRes.reduce((prevRetString, e) => `${e.regex},${prevRetString}`, '');
-    throw generateError("InvalidHarborConfiguration",
-      `We have multiple matching regexes for '${repoFullName}'`
-    );
-  } else if (matchingRes.length == 0 && !generateRegex(DEFAULT_REPO_DETAILS_MATCHER.regex).test(repoFullName)) {
-    throw generateError("HarborError",
-    `We have no matching regexes, including default, for '${repoFullName}'`
-    );
-  }
-
-  const matchPatternDetails = matchingRes.pop() || DEFAULT_REPO_DETAILS_MATCHER;
-  const {
-    lagoonProjectName = matchPatternDetails.defaultLagoonProject,
-    lagoonEnvironmentName = matchPatternDetails.defaultLagoonEnvironment,
-    lagoonServiceName = matchPatternDetails.defaultLagoonService,
-  } = extractRepositoryDetailsGivenRegex(repoFullName, matchPatternDetails.regex);
-
-  return {lagoonProjectName, lagoonEnvironmentName, lagoonServiceName};
-}
-
-const generateRegex = R.memoizeWith(R.identity, re => new RegExp(re));
-
-const extractRepositoryDetailsGivenRegex = (repoFullName, pattern = DEFAULT_REPO_DETAILS_REGEX) => {
-  const re = generateRegex(pattern);
-  const match = re.exec(repoFullName);
-  return match.groups || {};
-}
 
 const generateWebhookData = (
   webhookGiturl,
@@ -217,17 +137,7 @@ const generateWebhookData = (
   };
 };
 
-const extractVulnerabilities = (harborScanResponse) => {
-  for (let [key, value] of Object.entries(harborScanResponse)) {
-    let potentialStore: any = value;
-    if (potentialStore.hasOwnProperty('vulnerabilities')) {
-      return potentialStore.vulnerabilities;
-    }
-  }
-  throw new ProblemsHarborConnectionError(
-    "Scan response from Harbor does not contain a 'vulnerabilities' key"
-  );
-};
+
 
 const getVulnerabilitiesFromHarbor = async (scanId) => {
   let harborPayload = null;
@@ -241,18 +151,3 @@ const getVulnerabilitiesFromHarbor = async (scanId) => {
 
   return extractVulnerabilities(harborPayload);
 };
-
-class ProblemsHarborConnectionError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'problems-harborConnectionError';
-  }
-}
-
-class ProblemsInvalidWebhookData extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'problems-invalidWebhookData';
-  }
-}
-
